requirements.txt
astroid==3.3.7
black==24.10.0
click==8.1.8
dill==0.3.9
flake8==7.1.1
isort==5.13.2
mccabe==0.7.0
mypy==1.14.0
mypy-extensions==1.0.0
numpy==1.24.3
matplotlib==3.7.1
PyQt5==5.15.9
PyOpenGL==3.1.7
reportlab==4.0.4
scipy==1.10.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
pylint==3.3.2
ruff==0.8.4
tomlkit==0.13.2
typing_extensions==4.12.2
